refactor(furnacemind): rename AI cooperate modules
