+++ src/ui/furnacemind/chat_interface.py
